+++ api/src/main/java/dev/handsup/auction/controller/AuctionApiController.java
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import dev.handsup.auction.dto.ApiAuctionMapper;
-import dev.handsup.auction.dto.AuctionResponse;
+import dev.handsup.auction.dto.AuctionApiMapper;
+import dev.handsup.auction.dto.response.AuctionResponse;
 import dev.handsup.auction.dto.RegisterAuctionApiRequest;
-import dev.handsup.auction.dto.RegisterAuctionRequest;
+import dev.handsup.auction.dto.request.RegisterAuctionRequest;
 import dev.handsup.auction.service.AuctionService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -21,7 +21,7 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/auctions")
-public class AuctionController {
+public class AuctionApiController {
 
 	private final AuctionService auctionService;
 
@@ -29,7 +29,7 @@ public class AuctionController {
 	@ApiResponse(useReturnTypeSchema = true)
 	@PostMapping
 	public ResponseEntity<AuctionResponse> registerAuction(@Valid @RequestBody RegisterAuctionApiRequest request) {
-		RegisterAuctionRequest registerAuctionRequest = ApiAuctionMapper.toRegisterAuctionRequest(request);
+		RegisterAuctionRequest registerAuctionRequest = AuctionApiMapper.toRegisterAuctionRequest(request);
 		AuctionResponse response = auctionService.registerAuction(registerAuctionRequest);
 		return ResponseEntity.ok(response);
 	}
