+++ api/src/test/java/dev/handsup/auction/controller/AuctionApiControllerTest.java
@@ -16,8 +16,8 @@ import dev.handsup.auction.domain.auction_field.TradeMethod;
 import dev.handsup.auction.domain.product.ProductStatus;
 import dev.handsup.auction.domain.product.product_category.ProductCategory;
 import dev.handsup.auction.dto.RegisterAuctionApiRequest;
-import dev.handsup.auction.repository.AuctionRepository;
-import dev.handsup.auction.repository.ProductCategoryRepository;
+import dev.handsup.auction.repository.auction.AuctionRepository;
+import dev.handsup.auction.repository.product.ProductCategoryRepository;
 import dev.handsup.common.support.ApiTestSupport;
 import dev.handsup.fixture.ProductFixture;
