+++ core/src/main/java/dev/handsup/auction/dto/response/AuctionResponse.java
